+++ docs/changelog.rst
@@ -3,3 +3,4 @@ Changelog
 
 - Fix bug with a debug ``data`` file being created when using rtfm indexed presets
 - Fix bug with some presets failing to load on startup due to extra slashes at the end of the url.
+- Add the ability to import and export settings